Removed NativeNTSSPI and SafeNetHandles
