+++ src/Common/src/Interop/Windows/secur32/Interop.SSPI.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System;
+using System.Net.Security;
 using System.Runtime.InteropServices;
 
 internal static partial class Interop
@@ -58,6 +59,134 @@ internal static partial class Interop
               [In, Out] SecurityBufferDescriptor inputOutput,
               [In] uint sequenceNumber,
                    uint* qualityOfProtection
-              );        
+              );
+
+        [DllImport(Interop.Libraries.Secur32, ExactSpelling = true, SetLastError = true)]
+        internal static extern int QuerySecurityContextToken(
+            ref SSPIHandle phContext,
+            [Out] out SecurityContextTokenHandle handle);
+
+        [DllImport(Interop.Libraries.Secur32, ExactSpelling = true, SetLastError = true)]
+        internal static extern int FreeContextBuffer(
+            [In] IntPtr contextBuffer);
+
+        [DllImport(Interop.Libraries.Secur32, ExactSpelling = true, SetLastError = true)]
+        internal static extern int FreeCredentialsHandle(
+              ref SSPIHandle handlePtr
+              );
+
+        [DllImport(Interop.Libraries.Secur32, ExactSpelling = true, SetLastError = true)]
+        internal static extern int DeleteSecurityContext(
+              ref SSPIHandle handlePtr
+              );
+
+        [DllImport(Interop.Libraries.Secur32, ExactSpelling = true, SetLastError = true)]
+        internal unsafe static extern int AcceptSecurityContext(
+                  ref SSPIHandle credentialHandle,
+                  [In] void* inContextPtr,
+                  [In] SecurityBufferDescriptor inputBuffer,
+                  [In] ContextFlags inFlags,
+                  [In] Endianness endianness,
+                  ref SSPIHandle outContextPtr,
+                  [In, Out] SecurityBufferDescriptor outputBuffer,
+                  [In, Out] ref ContextFlags attributes,
+                  out long timeStamp
+                  );
+
+        [DllImport(Interop.Libraries.Secur32, ExactSpelling = true, SetLastError = true)]
+        internal unsafe static extern int QueryContextAttributesW(
+            ref SSPIHandle contextHandle,
+            [In] ContextAttribute attribute,
+            [In] void* buffer);
+
+        [DllImport(Interop.Libraries.Secur32, ExactSpelling = true, SetLastError = true)]
+        internal unsafe static extern int SetContextAttributesW(
+            ref SSPIHandle contextHandle,
+            [In] ContextAttribute attribute,
+            [In] byte[] buffer,
+            [In] int bufferSize);
+
+        [DllImport(Interop.Libraries.Secur32, ExactSpelling = true, SetLastError = true)]
+        internal static extern int EnumerateSecurityPackagesW(
+            [Out] out int pkgnum,
+            [Out] out SafeFreeContextBuffer_SECURITY handle);
+
+        [DllImport(Interop.Libraries.Secur32, ExactSpelling = true, CharSet = CharSet.Unicode, SetLastError = true)]
+        internal unsafe static extern int AcquireCredentialsHandleW(
+                  [In] string principal,
+                  [In] string moduleName,
+                  [In] int usage,
+                  [In] void* logonID,
+                  [In] ref AuthIdentity authdata,
+                  [In] void* keyCallback,
+                  [In] void* keyArgument,
+                  ref SSPIHandle handlePtr,
+                  [Out] out long timeStamp
+                  );
+
+        [DllImport(Interop.Libraries.Secur32, ExactSpelling = true, CharSet = CharSet.Unicode, SetLastError = true)]
+        internal unsafe static extern int AcquireCredentialsHandleW(
+                  [In] string principal,
+                  [In] string moduleName,
+                  [In] int usage,
+                  [In] void* logonID,
+                  [In] IntPtr zero,
+                  [In] void* keyCallback,
+                  [In] void* keyArgument,
+                  ref SSPIHandle handlePtr,
+                  [Out] out long timeStamp
+                  );
+
+        [DllImport(Interop.Libraries.Secur32, ExactSpelling = true, CharSet = CharSet.Unicode, SetLastError = true)]
+        internal unsafe static extern int AcquireCredentialsHandleW(
+                  [In] string principal,
+                  [In] string moduleName,
+                  [In] int usage,
+                  [In] void* logonID,
+                  [In] SafeSspiAuthDataHandle authdata,
+                  [In] void* keyCallback,
+                  [In] void* keyArgument,
+                  ref SSPIHandle handlePtr,
+                  [Out] out long timeStamp
+                  );
+
+        [DllImport(Interop.Libraries.Secur32, ExactSpelling = true, CharSet = CharSet.Unicode, SetLastError = true)]
+        internal unsafe static extern int AcquireCredentialsHandleW(
+                  [In] string principal,
+                  [In] string moduleName,
+                  [In] int usage,
+                  [In] void* logonID,
+                  [In] ref SecureCredential authData,
+                  [In] void* keyCallback,
+                  [In] void* keyArgument,
+                  ref SSPIHandle handlePtr,
+                  [Out] out long timeStamp
+                  );
+
+        [DllImport(Interop.Libraries.Secur32, ExactSpelling = true, SetLastError = true)]
+        internal unsafe static extern int InitializeSecurityContextW(
+                  ref SSPIHandle credentialHandle,
+                  [In] void* inContextPtr,
+                  [In] byte* targetName,
+                  [In] ContextFlags inFlags,
+                  [In] int reservedI,
+                  [In] Endianness endianness,
+                  [In] SecurityBufferDescriptor inputBuffer,
+                  [In] int reservedII,
+                  ref SSPIHandle outContextPtr,
+                  [In, Out] SecurityBufferDescriptor outputBuffer,
+                  [In, Out] ref ContextFlags attributes,
+                  out long timeStamp
+                  );
+
+        [DllImport(Interop.Libraries.Secur32, ExactSpelling = true, SetLastError = true)]
+        internal unsafe static extern int CompleteAuthToken(
+                  [In] void* inContextPtr,
+                  [In, Out] SecurityBufferDescriptor inputBuffers
+                  );
+
+        [DllImport(Interop.Libraries.Secur32, ExactSpelling = true, SetLastError = true)]
+        internal unsafe static extern SecurityStatus SspiFreeAuthIdentity(
+            [In] IntPtr authData);
     }
 }
