+++ src/System.Net.Security/src/Interop.TODO.cs
@@ -858,155 +858,4 @@ internal static partial class Interop
             return certificateProblems.ToArray();
         }
     }
-
-    internal static class SafeNetHandles
-    {
-        [DllImport(Interop.Libraries.Secur32, ExactSpelling = true, SetLastError = true)]
-        internal static extern int QuerySecurityContextToken(ref Interop.Secur32.SSPIHandle phContext, [Out] out SafeCloseHandle handle);
-
-        [DllImport(Interop.Libraries.Handle, ExactSpelling = true, SetLastError = true)]
-        internal static extern bool CloseHandle(IntPtr handle);
-
-        [DllImport(Interop.Libraries.Crypt32, ExactSpelling = true, SetLastError = true)]
-        internal static extern void CertFreeCertificateChain(
-            [In] IntPtr pChainContext);
-
-        [DllImport(Interop.Libraries.Crypt32, ExactSpelling = true, SetLastError = true)]
-        internal static extern void CertFreeCertificateChainList(
-            [In] IntPtr ppChainContext);
-
-        [DllImport(Interop.Libraries.Crypt32, ExactSpelling = true, SetLastError = true)]
-        internal static extern bool CertFreeCertificateContext(      // Suppressing returned status check, it's always==TRUE,
-            [In] IntPtr certContext);
-    }
-
-    internal static class SafeNetHandles_SECURITY
-    {
-        [DllImport(Interop.Libraries.Secur32, ExactSpelling = true, SetLastError = true)]
-        internal static extern int FreeContextBuffer(
-            [In] IntPtr contextBuffer);
-
-        [DllImport(Interop.Libraries.Secur32, ExactSpelling = true, SetLastError = true)]
-        internal static extern int FreeCredentialsHandle(
-              ref Interop.Secur32.SSPIHandle handlePtr
-              );
-
-        [DllImport(Interop.Libraries.Secur32, ExactSpelling = true, SetLastError = true)]
-        internal static extern int DeleteSecurityContext(
-              ref Interop.Secur32.SSPIHandle handlePtr
-              );
-
-        [DllImport(Interop.Libraries.Secur32, ExactSpelling = true, SetLastError = true)]
-        internal unsafe static extern int AcceptSecurityContext(
-                  ref Interop.Secur32.SSPIHandle credentialHandle,
-                  [In] void* inContextPtr,
-                  [In] SecurityBufferDescriptor inputBuffer,
-                  [In] ContextFlags inFlags,
-                  [In] Endianness endianness,
-                  ref Interop.Secur32.SSPIHandle outContextPtr,
-                  [In, Out] SecurityBufferDescriptor outputBuffer,
-                  [In, Out] ref ContextFlags attributes,
-                  out long timeStamp
-                  );
-
-        [DllImport(Interop.Libraries.Secur32, ExactSpelling = true, SetLastError = true)]
-        internal unsafe static extern int QueryContextAttributesW(
-            ref Interop.Secur32.SSPIHandle contextHandle,
-            [In] ContextAttribute attribute,
-            [In] void* buffer);
-
-        [DllImport(Interop.Libraries.Secur32, ExactSpelling = true, SetLastError = true)]
-        internal unsafe static extern int SetContextAttributesW(
-            ref Interop.Secur32.SSPIHandle contextHandle,
-            [In] ContextAttribute attribute,
-            [In] byte[] buffer,
-            [In] int bufferSize);
-
-        [DllImport(Interop.Libraries.Secur32, ExactSpelling = true, SetLastError = true)]
-        internal static extern int EnumerateSecurityPackagesW(
-            [Out] out int pkgnum,
-            [Out] out SafeFreeContextBuffer_SECURITY handle);
-
-        [DllImport(Interop.Libraries.Secur32, ExactSpelling = true, CharSet = CharSet.Unicode, SetLastError = true)]
-        internal unsafe static extern int AcquireCredentialsHandleW(
-                  [In] string principal,
-                  [In] string moduleName,
-                  [In] int usage,
-                  [In] void* logonID,
-                  [In] ref AuthIdentity authdata,
-                  [In] void* keyCallback,
-                  [In] void* keyArgument,
-                  ref Interop.Secur32.SSPIHandle handlePtr,
-                  [Out] out long timeStamp
-                  );
-
-        [DllImport(Interop.Libraries.Secur32, ExactSpelling = true, CharSet = CharSet.Unicode, SetLastError = true)]
-        internal unsafe static extern int AcquireCredentialsHandleW(
-                  [In] string principal,
-                  [In] string moduleName,
-                  [In] int usage,
-                  [In] void* logonID,
-                  [In] IntPtr zero,
-                  [In] void* keyCallback,
-                  [In] void* keyArgument,
-                  ref Interop.Secur32.SSPIHandle handlePtr,
-                  [Out] out long timeStamp
-                  );
-
-        //  Win7+
-        [DllImport(Interop.Libraries.Secur32, ExactSpelling = true, CharSet = CharSet.Unicode, SetLastError = true)]
-        internal unsafe static extern int AcquireCredentialsHandleW(
-                  [In] string principal,
-                  [In] string moduleName,
-                  [In] int usage,
-                  [In] void* logonID,
-                  [In] SafeSspiAuthDataHandle authdata,
-                  [In] void* keyCallback,
-                  [In] void* keyArgument,
-                  ref Interop.Secur32.SSPIHandle handlePtr,
-                  [Out] out long timeStamp
-                  );
-
-        [DllImport(Interop.Libraries.Secur32, ExactSpelling = true, CharSet = CharSet.Unicode, SetLastError = true)]
-        internal unsafe static extern int AcquireCredentialsHandleW(
-                  [In] string principal,
-                  [In] string moduleName,
-                  [In] int usage,
-                  [In] void* logonID,
-                  [In] ref SecureCredential authData,
-                  [In] void* keyCallback,
-                  [In] void* keyArgument,
-                  ref Interop.Secur32.SSPIHandle handlePtr,
-                  [Out] out long timeStamp
-                  );
-
-        [DllImport(Interop.Libraries.Secur32, ExactSpelling = true, SetLastError = true)]
-        internal unsafe static extern int InitializeSecurityContextW(
-                  ref Interop.Secur32.SSPIHandle credentialHandle,
-                  [In] void* inContextPtr,
-                  [In] byte* targetName,
-                  [In] ContextFlags inFlags,
-                  [In] int reservedI,
-                  [In] Endianness endianness,
-                  [In] SecurityBufferDescriptor inputBuffer,
-                  [In] int reservedII,
-                  ref Interop.Secur32.SSPIHandle outContextPtr,
-                  [In, Out] SecurityBufferDescriptor outputBuffer,
-                  [In, Out] ref ContextFlags attributes,
-                  out long timeStamp
-                  );
-
-        [DllImport(Interop.Libraries.Secur32, ExactSpelling = true, SetLastError = true)]
-        internal unsafe static extern int CompleteAuthToken(
-                  [In] void* inContextPtr,
-                  [In, Out] SecurityBufferDescriptor inputBuffers
-                  );
-    }
-
-    internal static class SspiHelper
-    {
-        [DllImport(Interop.Libraries.Secur32, ExactSpelling = true, SetLastError = true)]
-        internal unsafe static extern SecurityStatus SspiFreeAuthIdentity(
-            [In] IntPtr authData);
-    }
 }
