+++ src/System.Net.Security/src/System/Net/SecurityContextTokenHandle.cs
@@ -1,0 +1,33 @@
+// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using System.Threading;
+
+namespace System.Net.Security
+{
+#if DEBUG
+    internal sealed class SecurityContextTokenHandle : DebugCriticalHandleZeroOrMinusOneIsInvalid
+    {
+#else
+    internal sealed class SecurityContextTokenHandle : CriticalHandleZeroOrMinusOneIsInvalid
+    {
+#endif
+        private int _disposed;
+
+        private SecurityContextTokenHandle() : base()
+        {
+        }
+
+        protected override bool ReleaseHandle()
+        {
+            if (!IsInvalid)
+            {
+                if (Interlocked.Increment(ref _disposed) == 1)
+                {
+                    return Interop.mincore.CloseHandle(handle);
+                }
+            }
+            return true;
+        }
+    }
+}
